fix activity directive builder placement
