+++ src/routes/plans/[id]/+page.svelte
@@ -19,7 +19,6 @@
   import type { PaneAPI } from 'paneforge';
   import { onDestroy } from 'svelte';
   import { get } from 'svelte/store';
-  import ActivityDirectiveBuilder from '../../../components/activity/ActivityDirectiveBuilder.svelte';
   import Nav from '../../../components/app/Nav.svelte';
   import PageTitle from '../../../components/app/PageTitle.svelte';
   import Console from '../../../components/console/Console.svelte';
@@ -60,7 +59,6 @@
     resetPlanConstraintStores,
     uncheckedConstraintCount,
   } from '../../../stores/constraints';
-  import { directiveBuilderIsVisible, resetDirectiveBuilderStores } from '../../../stores/directiveBuilder';
   import {
     activityErrorRollups,
     allLogs,
@@ -106,7 +104,6 @@
     planSnapshotId,
     resetPlanSnapshotStores,
   } from '../../../stores/planSnapshots';
-  import { plugins } from '../../../stores/plugins';
   import {
     enableScheduling,
     latestSchedulingRequest,
@@ -144,7 +141,6 @@
     viewTogglePanel,
     viewUpdateGrid,
   } from '../../../stores/views';
-  import type { ActivityDirectiveInsertInput } from '../../../types/activity';
   import type { ActivityErrorCounts, LogLevel } from '../../../types/errors';
   import type { Extension } from '../../../types/extension';
   import type { PlanSnapshot } from '../../../types/plan-snapshot';
@@ -166,7 +162,6 @@
   } from '../../../utilities/simulation';
   import { getHumanReadableStatus, statusColors } from '../../../utilities/status';
   import { pluralize } from '../../../utilities/text';
-  import { formatDate } from '../../../utilities/time';
   import { showSuccessToast } from '../../../utilities/toast';
   import { tooltip } from '../../../utilities/tooltip';
   import { getSearchParameterNumber, removeQueryParam, setQueryParam } from '../../../utilities/url';
@@ -193,7 +188,6 @@
   };
   let compactNavMode = false;
   let constraintsStatusText: string | undefined;
-  let directiveBuilder: ActivityDirectiveBuilder;
   let hasCreateViewPermission: boolean = false;
   let hasUpdateViewPermission: boolean = false;
   let hasExpandPermission: boolean = false;
@@ -760,42 +754,11 @@
       }
     }
   }
-
-  async function onCreateActivityDirective(directive: ActivityDirectiveInsertInput) {
-    if ($plan !== null && $plan.model) {
-      // Convert offset to absolute start with plan as anchor
-      const offsetAsMs = getUnixEpochTimeFromInterval($plan.start_time, directive.start_offset);
-      const formattedStart = formatDate(new Date(offsetAsMs), $plugins.time.primary.format);
-      const newDirectiveId: number | null = await effects.createActivityDirective(
-        directive.arguments,
-        formattedStart,
-        directive.type,
-        directive.name,
-        directive.metadata,
-        $plan,
-        $user,
-      );
-      if (newDirectiveId !== null) {
-        $directiveBuilderIsVisible = false;
-        resetDirectiveBuilderStores();
-      }
-    }
-  }
 </script>
 
 <svelte:window on:keydown={onKeydown} bind:innerWidth={windowWidth} />
 
 <PageTitle subTitle={$plan?.name} title="Plans" />
-
-<ActivityDirectiveBuilder
-  bind:this={directiveBuilder}
-  width={200}
-  plan={$plan}
-  on:createActivityDirective={event => {
-    onCreateActivityDirective(event.detail.directive);
-  }}
-  user={$user}
-/>
 
 <div class="plan-container">
   <Resizable.PaneGroup direction="vertical" autoSaveId="console">
